+++ packages/draft-n-draw-react/src/index.tsx
@@ -1,1 +1,1 @@
-export * from './drafter';
+export { Drafter, getDrafter, useDrafter } from './drafter';