+++ packages/draft-n-draw-vanilla/src/index.ts
@@ -1,4 +1,4 @@
-export * from './drafter';
+export { Drafter } from './drafter';
 export * from './helpers/extended-box3-helper';
 export * from './helpers/point-helper';
 export * from './helpers/ray-helper';