+++ packages/draft-n-draw-react/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @draft-n-draw/react
 
+## 0.3.0
+
+### Minor Changes
+
+- Draws can now be made persistant. A dispose() method removes them from the scene.
+
+### Patch Changes
+
+- Updated dependencies
+  - @draft-n-draw/vanilla@0.3.0
+
+## 0.2.0
+
+### Minor Changes
+
+- Limit exports
+
+### Patch Changes
+
+- Updated dependencies
+  - @draft-n-draw/vanilla@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes