+++ packages/draft-n-draw-vanilla/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @draft-n-draw/vanilla
 
+## 0.3.0
+
+### Minor Changes
+
+- Draws can now be made persistant. A dispose() method removes them from the scene.
+
+## 0.2.0
+
+### Minor Changes
+
+- Limit exports
+
 ## 0.1.1
 
 ### Patch Changes